Complying with RFC5321 regarding MAIL FROM
The way mail_from is handled in the current code is non-compliant and non-working in case of server complying with the norm.
For example, you can't use a MS Exchange server with current code.

The "MAIL FROM:" was using either dynamic_username or dynamic_username_with_display_name (inconsistant behavior).
In both case the username field of server mail configuration was used. The username  field can be any string instead of an email address. In lots of scenario you don't use your email address as smtp username. In cas of exchange, it could be "username@domain" (which can be invalid as an email address) or "domain/username" which is also no a mail address.

To fix this we MUST use an email address in MAIL FROM, either in the form of "email@domain.tld" or "User Name <email@domain.tld>"

The commit fix this by ensuring a constant behavior among all code for sending email with the "email from" instead of "username" mail backend configuration.